+++ source/react/helpers/statics.js
@@ -1,0 +1,11 @@
+export function isNodeInRoot(node, root) {
+  while (node) {
+    if (node === root) {
+      return true;
+    }
+
+    node = node.parentNode;
+  }
+
+  return false;
+}