+++ source/react/library/menu/MenuList.js
@@ -3,7 +3,6 @@ import classnames from 'classnames';
 import MenuItem from './MenuItem';
 
 const propTypes = {
-  menuItemProps: React.PropTypes.object,
   options: React.PropTypes.array,
   selected: React.PropTypes.oneOfType([
     React.PropTypes.string,
@@ -15,7 +14,6 @@ const propTypes = {
 };
 
 const defaultProps = {
-  menuItemProps: {},
   selected: '',
 };
 
@@ -31,7 +29,7 @@ class MenuList extends React.Component {
   }
 
   render() {
-    const { selected, options, multiple, menuItemProps } = this.props;
+    const { selected, options, multiple } = this.props;
     const className = classnames('rc-menu-list', {
       'rc-menu-multiple': multiple,
     });
@@ -53,8 +51,6 @@ class MenuList extends React.Component {
           selected={ isSelected }
           onClick={ this.onChange }
           multiple={ multiple }
-          onMouseDown={ this.onChange }
-          { ...menuItemProps }
         />,
       );
     });