+++ source/react/library/select/Select.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import classnames from 'classnames';
 
+import { isNodeInRoot } from '../../helpers/statics';
+
 import Icon from '../Icon';
 import Input from '../Input';
 import Menu from '../menu/Menu';
@@ -61,6 +63,7 @@ class Select extends React.Component {
       options: formatOptions(props.options),
     };
 
+    this.onBlur = this.onBlur.bind(this);
     this.onSelect = this.onSelect.bind(this);
     this.onInputClick = this.onInputClick.bind(this);
   }
@@ -101,6 +104,13 @@ class Select extends React.Component {
     });
   }
 
+  onBlur(e) {
+    // If the user has clicked on a menu item, we handle that separately in onSelect.
+    if (!e.relatedTarget || !isNodeInRoot(e.relatedTarget, this.elem)) {
+      this.setState({ open: false });
+    }
+  }
+
   getCurrentValue() {
     let value = '';
 
@@ -119,9 +129,6 @@ class Select extends React.Component {
 
   renderMenuList() {
     let selected;
-    const menuItemProps = {
-      onClick: () => {},
-    };
 
     selected = this.state.options.filter(o => o.selected);
 
@@ -134,7 +141,6 @@ class Select extends React.Component {
         selected={ selected }
         options={ this.state.options }
         onChange={ this.onSelect }
-        menuItemProps={ menuItemProps }
       />
     );
   }
@@ -163,7 +169,7 @@ class Select extends React.Component {
         onChange={ e => this.setState({ inputValue: e.target.value }) }
         value={ this.getCurrentValue() }
         size={ this.props.size }
-        onBlur={ () => this.setState({ open: false }) }
+        onBlur={ this.onBlur }
         ref={ (c) => { this.input = c; } }
         focused={ this.state.open }
         onFocus={ () => this.setState({ open: true }) }
@@ -188,7 +194,7 @@ class Select extends React.Component {
     });
 
     return (
-      <div className={ className }>
+      <div ref={ (c) => { this.elem = c; } } className={ className }>
         { input }
         { menu }
       </div>